puts the elevator in elivator subsystems \
